Move server initializers up one module (#121)
* refactor: move DefaultStore implementation in block producer into store module

* refactor: move server initializers up one module

---------
