+++ rpc/src/main.rs
@@ -5,7 +5,7 @@ use cli::{Cli, Command, Request};
 use hex::ToHex;
 use miden_crypto::merkle::{path_to_text, TieredSmtProof};
 use miden_node_proto::{requests::CheckNullifiersRequest, rpc::api_client, tsmt::NullifierProof};
-use miden_node_rpc::{config::RpcTopLevelConfig, server::api};
+use miden_node_rpc::{config::RpcTopLevelConfig, server};
 use miden_node_utils::config::load_config;
 
 #[tokio::main]
@@ -18,7 +18,7 @@ async fn main() -> Result<()> {
 
     match cli.command {
         Command::Serve => {
-            api::serve(config.rpc).await?;
+            server::serve(config.rpc).await?;
         },
         Command::Request(req) => match req {
             Request::CheckNullifiers { nullifiers } => {
