+++ store/src/server/api.rs
@@ -1,4 +1,4 @@
-use std::{net::ToSocketAddrs, sync::Arc};
+use std::sync::Arc;
 
 use anyhow::Result;
 use miden_crypto::hash::rpo::RpoDigest;
@@ -16,41 +16,15 @@ use miden_node_proto::{
     },
     store::api_server,
 };
-use tonic::{transport::Server, Response, Status};
-use tracing::{info, instrument};
+use tonic::{Response, Status};
 
-use crate::{config::StoreConfig, db::Db, state::State, COMPONENT};
-
-// STORE INITIALIZER
-// ================================================================================================
-
-#[instrument(skip(config, db))]
-pub async fn serve(
-    config: StoreConfig,
-    db: Db,
-) -> Result<()> {
-    let endpoint = (config.endpoint.host.as_ref(), config.endpoint.port);
-    let addrs: Vec<_> = endpoint.to_socket_addrs()?.collect();
-
-    let state = Arc::new(State::load(db).await?);
-    let store = api_server::ApiServer::new(StoreApi { state });
-
-    info!(
-        host = config.endpoint.host,
-        port = config.endpoint.port,
-        COMPONENT,
-        "Server initialized",
-    );
-    Server::builder().add_service(store).serve(addrs[0]).await?;
-
-    Ok(())
-}
+use crate::state::State;
 
 // STORE API
 // ================================================================================================
 
 pub struct StoreApi {
-    state: Arc<State>,
+    pub(super) state: Arc<State>,
 }
 
 #[tonic::async_trait]
